Remove sdp interface (#2054)
* Remove SDP interface

* Remove SDP interface

* Add comments

* Add a new line in package.json

* rename Default SDP to SDP
