+++ docs/classes/audiovideocontrollerstate.html
@@ -419,7 +419,7 @@
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class">
 					<a name="previoussdpoffer" class="tsd-anchor"></a>
 					<h3>previous<wbr>Sdp<wbr>Offer</h3>
-					<div class="tsd-signature tsd-kind-icon">previous<wbr>Sdp<wbr>Offer<span class="tsd-signature-symbol">:</span> <a href="../interfaces/sdp.html" class="tsd-signature-type">SDP</a><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">null</span><span class="tsd-signature-symbol"> = null</span></div>
+					<div class="tsd-signature tsd-kind-icon">previous<wbr>Sdp<wbr>Offer<span class="tsd-signature-symbol">:</span> <a href="sdp.html" class="tsd-signature-type">SDP</a><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">null</span><span class="tsd-signature-symbol"> = null</span></div>
 					<aside class="tsd-sources">
 						<ul>
 							<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/main/src/audiovideocontroller/AudioVideoControllerState.ts#L45">src/audiovideocontroller/AudioVideoControllerState.ts:45</a></li>
