+++ docs/globals.html
@@ -173,7 +173,6 @@
 								<li class="tsd-kind-class"><a href="classes/defaultpingpong.html" class="tsd-kind-icon">Default<wbr>Ping<wbr>Pong</a></li>
 								<li class="tsd-kind-class"><a href="classes/defaultrealtimecontroller.html" class="tsd-kind-icon">Default<wbr>Realtime<wbr>Controller</a></li>
 								<li class="tsd-kind-class"><a href="classes/defaultreconnectcontroller.html" class="tsd-kind-icon">Default<wbr>Reconnect<wbr>Controller</a></li>
-								<li class="tsd-kind-class"><a href="classes/defaultsdp.html" class="tsd-kind-icon">DefaultSDP</a></li>
 								<li class="tsd-kind-class"><a href="classes/defaultsessionstatecontroller.html" class="tsd-kind-icon">Default<wbr>Session<wbr>State<wbr>Controller</a></li>
 								<li class="tsd-kind-class"><a href="classes/defaultsigv4.html" class="tsd-kind-icon">Default<wbr>Sig<wbr>V4</a></li>
 								<li class="tsd-kind-class"><a href="classes/defaultsignalingclient.html" class="tsd-kind-icon">Default<wbr>Signaling<wbr>Client</a></li>
@@ -258,6 +257,7 @@
 								<li class="tsd-kind-class"><a href="classes/reconnectionhealthpolicy.html" class="tsd-kind-icon">Reconnection<wbr>Health<wbr>Policy</a></li>
 								<li class="tsd-kind-class"><a href="classes/releasemediastreamstask.html" class="tsd-kind-icon">Release<wbr>Media<wbr>Streams<wbr>Task</a></li>
 								<li class="tsd-kind-class tsd-has-type-parameter"><a href="classes/runnabletask.html" class="tsd-kind-icon">Runnable<wbr>Task</a></li>
+								<li class="tsd-kind-class"><a href="classes/sdp.html" class="tsd-kind-icon">SDP</a></li>
 								<li class="tsd-kind-class"><a href="classes/sdpmediasection.html" class="tsd-kind-icon">SDPMedia<wbr>Section</a></li>
 								<li class="tsd-kind-class"><a href="classes/sendandreceivedatamessagestask.html" class="tsd-kind-icon">Send<wbr>And<wbr>Receive<wbr>Data<wbr>Messages<wbr>Task</a></li>
 								<li class="tsd-kind-class"><a href="classes/serialgrouptask.html" class="tsd-kind-icon">Serial<wbr>Group<wbr>Task</a></li>
@@ -384,7 +384,6 @@
 								<li class="tsd-kind-interface"><a href="interfaces/reconnectcontroller.html" class="tsd-kind-icon">Reconnect<wbr>Controller</a></li>
 								<li class="tsd-kind-interface"><a href="interfaces/removableanalysernode.html" class="tsd-kind-icon">Removable<wbr>Analyser<wbr>Node</a></li>
 								<li class="tsd-kind-interface"><a href="interfaces/removableobserver.html" class="tsd-kind-icon">Removable<wbr>Observer</a></li>
-								<li class="tsd-kind-interface"><a href="interfaces/sdp.html" class="tsd-kind-icon">SDP</a></li>
 								<li class="tsd-kind-interface"><a href="interfaces/scheduler.html" class="tsd-kind-icon">Scheduler</a></li>
 								<li class="tsd-kind-interface"><a href="interfaces/semver.html" class="tsd-kind-icon">Sem<wbr>Ver</a></li>
 								<li class="tsd-kind-interface"><a href="interfaces/sessionstatecontroller.html" class="tsd-kind-icon">Session<wbr>State<wbr>Controller</a></li>
@@ -1526,9 +1525,6 @@
 						<a href="classes/defaultreconnectcontroller.html" class="tsd-kind-icon">Default<wbr>Reconnect<wbr>Controller</a>
 					</li>
 					<li class=" tsd-kind-class">
-						<a href="classes/defaultsdp.html" class="tsd-kind-icon">DefaultSDP</a>
-					</li>
-					<li class=" tsd-kind-class">
 						<a href="classes/defaultsessionstatecontroller.html" class="tsd-kind-icon">Default<wbr>Session<wbr>State<wbr>Controller</a>
 					</li>
 					<li class=" tsd-kind-class">
@@ -1779,6 +1775,9 @@
 					</li>
 					<li class=" tsd-kind-class tsd-has-type-parameter">
 						<a href="classes/runnabletask.html" class="tsd-kind-icon">Runnable<wbr>Task</a>
+					</li>
+					<li class=" tsd-kind-class">
+						<a href="classes/sdp.html" class="tsd-kind-icon">SDP</a>
 					</li>
 					<li class=" tsd-kind-class">
 						<a href="classes/sdpmediasection.html" class="tsd-kind-icon">SDPMedia<wbr>Section</a>
@@ -2142,9 +2141,6 @@
 					</li>
 					<li class=" tsd-kind-interface">
 						<a href="interfaces/removableobserver.html" class="tsd-kind-icon">Removable<wbr>Observer</a>
-					</li>
-					<li class=" tsd-kind-interface">
-						<a href="interfaces/sdp.html" class="tsd-kind-icon">SDP</a>
 					</li>
 					<li class=" tsd-kind-interface">
 						<a href="interfaces/scheduler.html" class="tsd-kind-icon">Scheduler</a>
