+++ docs/index.html
@@ -1456,9 +1456,6 @@ appVersion: string;
 						<a href="classes/defaultreconnectcontroller.html" class="tsd-kind-icon">Default<wbr>Reconnect<wbr>Controller</a>
 					</li>
 					<li class=" tsd-kind-class">
-						<a href="classes/defaultsdp.html" class="tsd-kind-icon">DefaultSDP</a>
-					</li>
-					<li class=" tsd-kind-class">
 						<a href="classes/defaultsessionstatecontroller.html" class="tsd-kind-icon">Default<wbr>Session<wbr>State<wbr>Controller</a>
 					</li>
 					<li class=" tsd-kind-class">
@@ -1709,6 +1706,9 @@ appVersion: string;
 					</li>
 					<li class=" tsd-kind-class tsd-has-type-parameter">
 						<a href="classes/runnabletask.html" class="tsd-kind-icon">Runnable<wbr>Task</a>
+					</li>
+					<li class=" tsd-kind-class">
+						<a href="classes/sdp.html" class="tsd-kind-icon">SDP</a>
 					</li>
 					<li class=" tsd-kind-class">
 						<a href="classes/sdpmediasection.html" class="tsd-kind-icon">SDPMedia<wbr>Section</a>
@@ -2072,9 +2072,6 @@ appVersion: string;
 					</li>
 					<li class=" tsd-kind-interface">
 						<a href="interfaces/removableobserver.html" class="tsd-kind-icon">Removable<wbr>Observer</a>
-					</li>
-					<li class=" tsd-kind-interface">
-						<a href="interfaces/sdp.html" class="tsd-kind-icon">SDP</a>
 					</li>
 					<li class=" tsd-kind-interface">
 						<a href="interfaces/scheduler.html" class="tsd-kind-icon">Scheduler</a>
