+++ package.json
@@ -150,3 +150,4 @@
     "printWidth": 100
   }
 }
+
